Refactor, fix imports
