+++ ts/packages/knowPro/test/search.spec.ts
@@ -5,11 +5,9 @@ import {
     IConversation,
     KnowledgeType,
     SearchTermGroup,
-} from "../src/interfaces.js";
-import {
-    searchConversationKnowledge,
     SemanticRefSearchResult,
-} from "../src/search.js";
+} from "../src/interfaces.js";
+import { searchConversationKnowledge } from "../src/search.js";
 import {
     createSearchTerm,
     createAndTermGroup,
